Lambda Stream Studies

+++ src/Class2/WordCounter.java
@@ -1,6 +1,10 @@
 package Class2;
 
+import java.beans.Expression;
+import java.lang.reflect.Array;
+import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.List;
 
 public class WordCounter {
     public static void main(String[] args) {
@@ -9,11 +13,18 @@ public class WordCounter {
         System.out.println(counterOfWord(repeatedWords));
     }
 
-    public static int counterOfWord(String str) {
-        return (int) Arrays.stream(str.split(" "))
+    public static long counterOfWord(String str) {
+        return Arrays.stream(str.split(" "))
                 .filter(x -> x.equals("watch"))  //.contains("watch")
                 .count();
     }
 }
 
 
+//class testX {
+//    public static long methodX(String  str) {
+//        return Arrays.stream(str.split(" "))
+//                .filter(x -> x.equals("ll"))
+//                .count();
+//    }
+//}
